requirements.txt
python-telegram-bot
# openai
python-dotenv